[ARM-292] Force refresh the invitation creators telemetry after friend accepts it
