+++ rust/src/constants.rs
@@ -17,3 +17,7 @@ pub const ASIMOV_LIVES: &str = "asimovlives";
 pub static DATE_FORMAT: &str = "%Y-%m-%dT%H:%M:%S%.3fZ";
 
 pub static DEFAULT_NOTIFICATION_ICON: &str = "ic_stat_logo";
+
+pub static NOTIFICATIONS_EXCHANGE: &str = "notifications.exchange";
+
+pub static NOTIFICATIONS_ROUTING_KEY: &str = "notifications";
